oops, 2nd arg of getnameinfo() is socklen_t, not size_t. (this was changed somewhere before rfc2553)

+++ freebsd3/include/netdb.h
@@ -250,7 +250,7 @@ void		setnetent __P((int));
 void		setprotoent __P((int));
 int		getaddrinfo __P((const char *, const char *,
 				 const struct addrinfo *, struct addrinfo **));
-int		getnameinfo __P((const struct sockaddr *, size_t, char *,
+int		getnameinfo __P((const struct sockaddr *, unsigned int, char *,
 				 size_t, char *, size_t, int));
 void		freeaddrinfo __P((struct addrinfo *));
 char		*gai_strerror __P((int));
